Fix cpu hog of contrail-vrouter-agent caused by setting flow_cache_timeout to high values.
The high value of flow_cache_timeout when used to compute the flow timer interval, does not
fit in uint32 field. We need to interpret the value as uint64 instead of uint32.
Closes-Bug: #1505615

+++ src/vnsw/agent/vrouter/flow_stats/flow_stats_collector.cc
@@ -42,7 +42,7 @@ FlowStatsCollector::FlowStatsCollector(boost::asio::io_service &io, int intvl,
         flow_default_interval_ = intvl;
         if (flow_cache_timeout) {
             // Convert to usec
-            flow_age_time_intvl_ = 1000000 * flow_cache_timeout;
+            flow_age_time_intvl_ = 1000000L * (uint64_t)flow_cache_timeout;
         } else {
             flow_age_time_intvl_ = FlowAgeTime;
         }
@@ -58,7 +58,7 @@ void FlowStatsCollector::Shutdown() {
 }
 
 void FlowStatsCollector::UpdateFlowMultiplier() {
-    uint32_t age_time_millisec = flow_age_time_intvl_ / 1000;
+    uint64_t age_time_millisec = flow_age_time_intvl_ / 1000;
     if (age_time_millisec == 0) {
         age_time_millisec = 1;
     }
